mvp of concat page
